Use `ref:emptyToNull` function to handle empty `wikipedia_url` values.
Motivation:

ROR uses a mixture of `null` and empty string (`""`) values to express
that a particular organisation has no corresponding Wikipedia page.

RMLMapper skips any `null` values, but tried to map an empty string
value to a corresponding IRI.  This is undesirable because the empty
string represents missing data, which is better represented by not
including a triple.

In fact, mapping the empty string also causes RMLMapper to abort, beause
the empty string is not a valid IRI.

Modification:

Use the ref:emptyToNull function to consider empty strings as equivalent
to a null value.

Result:

Better handling of `wikipedia_url` values that are an empty string.