Update download script
The name of the json file has changed, breaking the assumption
underlying how the download script worked.

Update the script so that, rather than guessing the name of the json
file (contained within the zip file), it now queries the zip file for
this information.

This means we must always download the file.