Factor out redirection handling
Motivation:

Websites are unreliable at returning the correct redirections.  Some
sites return a temporary redirect (instead of a permanent redirect) when
upgrading from http:// to https://.  Similarly, some sites return a
permanent redirection when injecting a session ID.

Modification:

Include a function for processing redirections (if one was returned).
This includes heuristics to handle bad websites.

Result:

More reliable information on when a URL should be updated to a new URL.