Update redirection veto logic
Motivation:

Currently any redirection that includes a session key is vetoed.
However, the redirection could be a simple http:// --> https:// upgrade,
which should be accepted.

Modification:

Change heuristics to reject a permanent redirection only if this adds a
session.  If the targeted URL already contains a session then accept the
redirection.  Log such cases if they are not a simple scheme-only
change.

Result:

Redirections of URLs that already contain a session key are now
accepted.