Update output report
Motivation:

Update report format to be friendlier and to support multiple problems
with a single link.  This is in preparation for checking redirections.

Modification:

Update JSON output to include a label that describes the link in a
URL-agnostic fashion.  The value of the label is a JSON array of checks.
For each problem, there is the target URL, a generic problem class, a
more human-friendly description of what went wrong.  For redirections,
the new URL is provided as an extra field.

Result:

Output format has been updated to be more flexible and to provide a
better description of what went wrong.