Also check redirection URLs
Motivation:

Just because the server redirected to a new URL doesn't mean that this
new URL is good.  There could be further redirections, or the new URL
could be broken.

Modification:

Schedule a check for redirection target URL.

Update how futures are handled to avoid racy problems.

Result:

Stronger verification that the URL is good.