Add confirmation that redirection link is good
Motivation:

When accepting a redirection, the code schedules a check on the
updated/preferred URL.  If the link is good then there is nothing in the
report to confirm this, which may be confusing.

Modification:

Add ability to record a log a successful URL check.

Update redirection checks to record a successful URL check.

Result:

Redirections now have an extra line indicating the URL was good.

In the future, we may want to make this configurable, perhaps also
allowing the report to include such positive entries for regular
(non-redirected) URL checks.