Identify connection errors due to TLS
Motivation:

Bad web-sites have a problmatic certificate; however, they could still
respond with a redirection to a valid URL.

Modification:

As a first step, identify which failures are due to a TLS error.
Unfortunately, we cannot distinguish between certificate-related errors
and other TLS errors: urllib3 looses this information.

(The only solution here would be to parse the error message.)

Result:

The description is slightly modified to indicate the nature of the
problem.  This is redundant, but is used to demonstrate correct
behaviour.