Switch off certificate validation warnings
Motivation:

urllib3 logs (at WARNING level) if the integrety of a TLS connection
could not be validated.  In this specific case, this warning is
triggered if the remote site presents a certificate from a CA that we
don't trust.

While this warning is valid and (in general) a good idea, the same
information is captured in the report.  Therefore, logging a warning is
redundant and clutters up the output.

Modification:

Suppress WARNING-level logging for connections.

Result:

Validation output is less cluttered and easier to read.