Switch off warning that connection has not been verified
Motivation:

urllib3 will emit a warning if it makes a network connection that is
insecure.

In general, this is a good idea; however, the validate script does this
under specific conditions, in order to see if the server redirects to a
URL that "works".

Therefore, the warning is redundant and clutters up the output.

Modification:

Switch off warning that an https connection is insecure.

Result:

Less cluttered console output from script.