Add a shot-in-the-dark guess as a work-around
Motivation:

Some web-sites DNS resolve for a FQDN for which the host certificate has
no SAN, and for which there is no redirection rule to a FQDN for which
the host certificate has a SAN.

For example: 'www.example.org' and 'example.org' both DNS-resolve to the
same IP address, the server has a host certificate for 'example.org' but
not 'www.example.org', and there is no redirect rule from
'www.example.org' to 'example.org'.

Early web servers used the 'www' hostname (e.g., 'www.example.org');
however, the trend is now to drop this 'www' hostname (e.g.,
'example.org').

Therefore, if a URL has a hostname starting 'www.' and it is suffering
from a bad host certificate then dropping the 'www.' hostname from the
URL's FQDN is guess that might just work.

Modification:

Try removing the 'www.' hostname if there is a TLS problem and the
server returns a 200 status code with TLS host verification disabled
(i.e., no redirection).

Result:

Able to resolve more bad URLs.