Add redirection loop detection
Motivation:

Following redirection and guessing possible alternative hostnames can
lead to the code testing a URL a second time.  Beyond that this is
inefficient, the risk is that the code is caught in a loop and will
continue testing the URLs in this loop without ending.

Modification:

Keep track of the URLs processed per ID-label.

Check for a loop before submitting HEAD request.  Abort the request if a
loop is detected.

Add code to clean-up the URL history once an ID-label is no longer being
tested.

Result:

Code no longer gets caught in URL loops.