Don't terminate application while the call-back is active
Motivation:

The final callback may still wish to schedule more HTTP HEAD tests.
This requires the executor (thread pool) is still active and available.

The thread pool is automatically terminated once the application
exits.  Once this happens, no further tasks may be submitted.

Therefore, in order to support a final call-back that wishes to submit
further activity, the application must not exit while the call-back is
active.

Modification:

Add global boolean to record whether the callback is active.

Update call-back to keep this global boolean up-to-date.

Update main/wait loop to check whether the callback is active and
continue waiting if it is.

Result:

The application no longer terminates when the callback is active.