Add work-around for broken `verify` behaviour on session.
Motivation:

The `verify` kwarg to the `Session.head` method is not always honoured.
In particular, if the session has already established a TLS connection
to the server from an earlier call with `verify=False` then the new call
will reuse that connection.

The result is that an HTTP HEAD request may not report certificate
problems, even if `verify=True` is specified.

This has been reported as:

https://github.com/psf/requests/issues/6269

Modification:

Introduce a work-around: maintain a seperate Session for non-verified
connections.

Result:

HTTP requests with TLS verification enabled will only be made with an
existing connection if that existing connection was established with TLS
verification enabled.