Revamp retry strategy on TLS errors
Motivation:

The code currently retries with certificate verification disabled in
situations where this will not help (e.g., DH too small, bad TLS
version, and other non-certificate based TLS problems).

Additionally, the code always attempts to guess alternative FQDNs
(stripping off 'www.') if the URL was attempted with certificate
verification disabled.

If (for example) the certificate was issued by an untrusted CA then
guessing an alternative FQDN is unlikely to succeed.

Speed up test run by avoiding unnecessary URL tests.

Modification:

Add a boolean flag to indicate whether it makes sense to try a URL with
a modified FQDN.

Update test whether to retry with certificate verification disabled to
base that decision on whether the problem was due to a bad certificate.

Add method to populate this flag based on whether the problem was
because of the certificate SANs didn't match the FQDN.

Result:

Tests that cannot succeed are avoided.

If the certificate is issued by