Update progress report to include deltas
Motivation:

Printing number of completed URLs and total URLs tot check provide basic
feedback on progress.  However, it is hard to understand the change (the
progress made) since the previous report when the numbers become large.

Modification:

Calculate and show the delta from previous report.

Result:

It is easier to see what progress is being made.