Fix redirection URL in report
Motivation:

Currently the report logs the 'location' response header.  Although this
works in most cases, it is wrong as the server may respond with a
relative URL, which should be resolved against the request URL to obtain
the redirection.

Modification:

Record the resolved, absolute URL if the server responds with a
redirection.

Result:

The report no longer records relative URLs, even if the server's
'location' response header value is a relative URL.