Use a per-id-label cookie jar
Motivation:

Some web servers return cookies when they receive a HEAD request.  In
some cases, the response is a redirection that requires a subsequent
request to include previously cookies.

The current behaviour uses a session cookiejar.  This is bad because:

 1. we may make multiple requests to the same domain.  Storing cookies
    within the session risks the results from one (id,label) interfering
    with a different (id,label).

 2. over requests, the number of stored cookies will increase.  This
    can have a negative impact on performance.

 3. we have two sessions (a secure one and insecure one) as a
    work-around for bad bahaviour with the verify option.  However,
    these two sessions do not share cookies.

Modification:

Add a CookiePolicy that rejects all cookies.  Update the two session's
cookie jars to support this policy.

Generate a cookiejar per (id,label).  Responses that include cookies are
added to this cookiejar and requests include cookies from this jar.  The
jar is disposed of, once the (id,label) is finished.

Result:

Cookies are stored in seperate cookiejars, one per (id,label).