Consider cookies when checking for a loop
Motivation:

Some web sites use redirection to the same URL as a way to inject
cookies.  Currently the code considers this a loop and rejects the URL.

Modification:

Consider the contents of the (id,label) cookiejar when looking for
loops.  A request is now considers a loop only if it has the same
verification, same canonical URL and the same cookie contents.

Result:

The code now accepts web pages that use self-redirection as a way to
inject cookies into the browser.