Remove canonicalising loop check, add check for too many redirections
Motivation:

The canonical form for URLs considers a URL with a base path and one
without as equivalent; however, this is not true and causes problems.

Modification:

Remove code that stores a canonical form of URL.

Add an excessive redirection check, as a safety measure.

Refactor code and rename loop check as the now more generic pre-flight
check method.

Result:

Certain redirections (URL with '/' as the path to URL without a path)
are no longer considered equivalent when checking for loops.

An additional safety has been added, so any (id,label) cannot result in
too many checks.